src/main/java/com/ysx/leetcode/medium/LeetCode325.java
/*
 * Copyright (c) ysx. 2020-2020. All rights reserved.
 */

package com.ysx.leetcode.medium;

/**
 * @author youngbear
 * @email youngbear@aliyun.com
 * @date 2020/12/6 16:56
 * @blog https://blog.csdn.net/next_second
 * @github https://github.com/YoungBear
 * @description 325. 和等于 k 的最长子数组长度
 */
public class LeetCode325 {
    /**
     * 暴力法
     * 时间复杂度O(n*n)
     *
     * @param nums
     * @param k
     * @return
     */
    public int maxSubArrayLen(int[] nums, int k) {
        if (nums == null || nums.length == 0) {
            return 0;
        }
        int max = 0;
        for (int i = 0; i < nums.length; i++) {
            int sum = 0;
            for (int j = i; j < nums.length; j++) {
                sum += nums[j];
                if (sum == k) {
                    max = Math.max(max, j - i + 1);
                }
            }
        }
        return max;
    }
}
